requirements.txt
streamlit==1.32.0
openai
tavily-python==0.2.7
python-dotenv==1.0.0
